demos/dashboard/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
-e file:///Users/james/code/wildfish/django-datorum/django-datorum
    # via -r requirements.in
alabaster==0.7.12
    # via sphinx
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
appnope==0.1.3
    # via ipython
asgiref==3.5.2
    # via django
asttokens==2.0.8
    # via stack-data
attrs==22.1.0
    # via
    #   pytest
    #   pytest-mypy
babel==2.10.3
    # via sphinx
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via webtest
black==22.10.0
    # via pytest-black
build==0.8.0
    # via pip-tools
certifi==2022.9.24
    # via
    #   httpcore
    #   httpx
    #   requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
    #   strawberry-graphql
    #   uvicorn
coverage[toml]==6.5.0
    # via pytest-cov
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
distlib==0.3.6
    # via virtualenv
django==4.1.2
    # via
    #   django-cors-headers
    #   django-datorum
    #   django-grip
    #   django-htmx
    #   model-bakery
django-cors-headers==3.13.0
    # via
    #   -r ../../requirements.in
    #   django-datorum
django-eventstream==4.4.1
    # via -r requirements.in
django-grip==3.2.0
    # via django-eventstream
django-htmx==1.12.2
    # via
    #   -r ../../requirements.in
    #   django-datorum
django-webtest==1.9.10
    # via -r ../../requirements.in
docutils==0.17.1
    # via
    #   sphinx
    #   sphinx-rtd-theme
exceptiongroup==1.0.0rc9
    # via pytest
executing==1.1.1
    # via stack-data
faker==15.1.1
    # via -r ../../requirements.in
fastdiff==0.3.0
    # via snapshottest
filelock==3.8.0
    # via
    #   pytest-mypy
    #   virtualenv
flake8==5.0.4
    # via pytest-flake8
graphql-core==3.2.3
    # via strawberry-graphql
gripcontrol==4.1.0
    # via
    #   django-eventstream
    #   django-grip
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via -r requirements.in
identify==2.5.7
    # via pre-commit
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
imagesize==1.4.1
    # via sphinx
iniconfig==1.1.1
    # via pytest
ipdb==0.13.9
    # via -r ../../requirements.in
ipython==8.5.0
    # via
    #   -r ../../requirements.in
    #   ipdb
isort==5.10.1
    # via pytest-isort
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via sphinx
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via flake8
model-bakery==1.9.0
    # via -r ../../requirements.in
mypy==0.982
    # via pytest-mypy
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
nodeenv==1.7.0
    # via pre-commit
packaging==21.3
    # via
    #   build
    #   pytest
    #   sphinx
parso==0.8.3
    # via jedi
pathspec==0.10.1
    # via black
pep517==0.13.0
    # via build
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.9.0
    # via -r ../../requirements.in
platformdirs==2.5.2
    # via
    #   black
    #   virtualenv
plotly==5.10.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
pre-commit==2.20.0
    # via -r ../../requirements.in
prompt-toolkit==3.0.31
    # via ipython
psutil==5.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pubcontrol==3.3.0
    # via
    #   django-grip
    #   gripcontrol
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.9.1
    # via flake8
pydantic==1.10.2
    # via -r ../../requirements.in
pyflakes==2.5.0
    # via flake8
pygments==2.13.0
    # via
    #   ipython
    #   sphinx
    #   strawberry-graphql
pyjwt==2.6.0
    # via
    #   django-eventstream
    #   gripcontrol
    #   pubcontrol
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via
    #   -r ../../requirements.in
    #   pytest-black
    #   pytest-cov
    #   pytest-django
    #   pytest-flake8
    #   pytest-isort
    #   pytest-mypy
pytest-black==0.3.12
    # via -r ../../requirements.in
pytest-cov==4.0.0
    # via -r ../../requirements.in
pytest-django==4.5.2
    # via -r ../../requirements.in
pytest-flake8==1.1.1
    # via -r ../../requirements.in
pytest-isort==3.0.0
    # via -r ../../requirements.in
pytest-mypy==0.10.0
    # via -r ../../requirements.in
python-dateutil==2.8.2
    # via
    #   faker
    #   strawberry-graphql
python-multipart==0.0.5
    # via strawberry-graphql
pytz==2022.5
    # via babel
pyyaml==6.0
    # via pre-commit
requests==2.28.1
    # via
    #   pubcontrol
    #   sphinx
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via
    #   asttokens
    #   django-eventstream
    #   django-grip
    #   gripcontrol
    #   python-dateutil
    #   python-multipart
    #   snapshottest
snapshottest==0.6.0
    # via -r ../../requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.3.2.post1
    # via beautifulsoup4
sphinx==5.3.0
    # via
    #   -r ../../requirements.in
    #   sphinx-rtd-theme
sphinx-rtd-theme==1.0.0
    # via -r ../../requirements.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.3
    # via django
stack-data==0.5.1
    # via ipython
starlette==0.21.0
    # via strawberry-graphql
strawberry-graphql[debug-server]==0.137.1
    # via
    #   -r ../../requirements.in
    #   django-datorum
tenacity==8.1.0
    # via plotly
termcolor==2.0.1
    # via snapshottest
toml==0.10.2
    # via
    #   ipdb
    #   pre-commit
    #   pytest-black
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   mypy
    #   pep517
    #   pytest
traitlets==5.5.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.4.0
    # via
    #   mypy
    #   pydantic
    #   strawberry-graphql
urllib3==1.26.12
    # via requests
uvicorn==0.19.0
    # via strawberry-graphql
virtualenv==20.16.5
    # via pre-commit
waitress==2.1.2
    # via webtest
wasmer==1.1.0
    # via fastdiff
wasmer-compiler-cranelift==1.1.0
    # via fastdiff
wcwidth==0.2.5
    # via prompt-toolkit
webob==1.8.7
    # via webtest
webtest==3.0.0
    # via django-webtest
werkzeug==2.2.2
    # via django-grip
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
